refactor: organize parser into modular components
Split the large `parser.rs` file into logical submodules within
`front/parser/src/parser/` to improve code structure and maintainability.

Changes:
- **New Module Structure**:
  - `asm.rs`: `parse_asm_block`
  - `control.rs`: `parse_if`, `parse_while`, `parse_for`
  - `decl.rs`: `parse_variable_decl`, `parse_const`, `parse_let`, `parse_var`
  - `expr.rs`: `parse_function_call`, `parse_parentheses`
  - `functions.rs`: `parse_function`, `parse_parameters`, `extract_body`
  - `io.rs`: `parse_println`, `parse_print`, `parse_input`
  - `items.rs`: `parse_import`, `parse_proto`, `parse_struct`
  - `stmt.rs`: `parse_assignment`, `parse_block`, `parse_statement`
  - `types.rs`: Type parsing logic moved from `type_system.rs` (`parse_type`, `parse_type_from_token`, etc.)
  - `parse.rs`: Main `parse()` entry point and module declarations.
- **Cleanup**:
  - Removed `type_system.rs` (logic moved to `types.rs`).
  - Updated `mod.rs` to expose the new structure.
  - Fixed imports across the crate to point to new module locations.
  - Resolved unused import warnings in `main.rs` and `runner.rs`.

This refactoring decouples parser components, making it easier to navigate
and extend individual language features.
