+++ front/parser/src/lib.rs
@@ -1,3 +1,9 @@
 pub mod parser;
+pub mod ast;
+pub mod format;
+pub mod import;
+pub mod stdlib;
+pub mod type_system;
+pub mod verification;
 
 pub use parser::*;
